Move logic to its own file
